requirements.txt
fastapi
uvicorn[standard]
python-multipart
cryptography
PyJWT
pydantic
python-dotenv
google-generativeai
requests
streamlit
